minor edits

+++ src/SparkFunESP8266WiFi.h
@@ -111,8 +111,8 @@ private:
 	// Command Send/Receive //
 	//////////////////////////
 	void sendCommand(const char * cmd, enum esp8266_command_type type = ESP8266_CMD_EXECUTE, const char * params = NULL);
-	int16_t readForResponse(const char * rsp, unsigned int timeout);
-	int16_t readForResponses(const char * pass, const char * fail, unsigned int timeout);
+	int16_t readForResponse(const char * rsp, unsigned long timeout);
+	int16_t readForResponses(const char * pass, const char * fail, unsigned long timeout);
 	
 	//////////////////
 	// Buffer Stuff //
